+++ egui/src/containers/collapsing_header.rs
@@ -109,11 +109,9 @@ pub fn paint_icon(ui: &mut Ui, openness: f32, response: &Response) {
     // Draw a pointy triangle arrow:
     let rect = Rect::from_center_size(rect.center(), vec2(rect.width(), rect.height()) * 0.75);
     let mut points = vec![rect.left_top(), rect.right_top(), rect.center_bottom()];
-    let rotation = Vec2::angled(remap(openness, 0.0..=1.0, -TAU / 4.0..=0.0));
+    let rotation = Rot2::from_angle(remap(openness, 0.0..=1.0, -TAU / 4.0..=0.0));
     for p in &mut points {
-        let v = *p - rect.center();
-        let v = rotation.rotate_other(v);
-        *p = rect.center() + v;
+        *p = rect.center() + rotation * (*p - rect.center());
     }
 
     ui.painter().add(PaintCmd::closed_line(points, stroke));